Split NovoToast and ToastService
