+++ demo/pages/toast/_ToastDemo.scss
@@ -12,6 +12,18 @@ toast-demo {
     //    }
     //}
 
+    // @TODO @asibilia - reimplement novo-forms
+
+    .fake-card {
+        border-radius: 3px;
+        box-shadow: 0 1px 3px rgba(0,0,0,.2);
+        background: #fff;
+
+        .content {
+            padding: 20px;
+        }
+    }
+
     form {
         > * {
             display: block;
