+++ demo/pages/toast/templates/ToastDemo.html
@@ -1,5 +1,24 @@
-<button theme="dialogue" color="success" icon="coffee" (click)="toastToggled('top')"
-    data-automation-id="toast-trigger">Fixed Top</button>
-<button theme="dialogue" color="primary" icon="check" (click)="toastToggled('bottom')">Fixed Bottom</button>
-<button theme="dialogue" color="negative" icon="times" (click)="toastToggled('growlRight')">Growl: Top Right</button>
-<button theme="dialogue" color="primary" icon="coffee" (click)="toastToggled('growlLeft')">Growl: Top Left</button>
+<div class="fake-card">
+    <header title="John Deere" icon="person" theme="contact">
+        <utils>
+            <util-action icon="flag"></util-action>
+            <util-action icon="refresh"></util-action>
+            <util-action icon="times"></util-action>
+        </utils>
+    </header>
+    <novo-toast theme="danger" title="Save Failed" message="Oops! Looks like you're missing some required fields"
+        icon="caution" position="growlTopRight"></novo-toast>
+    <div class="content">
+        <p>
+            Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed
+            do eiusmod tempor incididunt ut labore et dolore
+            magna aliqua. Ut enim ad minim veniam, quis nostrud
+            exercitation ullamco laboris nisi ut aliquip ex ea
+            commodo consequat. Duis aute irure dolor in reprehenderit
+            in voluptate velit esse cillum dolore eu fugiat nulla
+            pariatur. Excepteur sint occaecat cupidatat non proident,
+            sunt in culpa qui officia deserunt mollit anim id
+            est laborum.
+        </p>
+    </div>
+</div>
