+++ demo/pages/toast/templates/ToastServiceDemo.html
@@ -1,0 +1,5 @@
+<button theme="dialogue" color="success" icon="coffee" (click)="toastToggled('top')"
+    data-automation-id="toast-trigger">Fixed Top</button>
+<button theme="dialogue" color="primary" icon="check" (click)="toastToggled('bottom')">Fixed Bottom</button>
+<button theme="dialogue" color="negative" icon="times" (click)="toastToggled('growlRight')">Growl: Top Right</button>
+<button theme="dialogue" color="primary" icon="coffee" (click)="toastToggled('growlLeft')">Growl: Top Left</button>
