+++ src/elements/toast.js
@@ -1,2 +1,3 @@
 export { NOVO_TOAST_ELEMENTS } from './toast/Toast';
-export { NovoToast, NovoToaster } from './toast/Toast';
+export { NovoToast } from './toast/Toast';
+export { ToastService } from './toast/ToastService';
