+++ src/elements/toast/Toast.e2e.js
@@ -13,4 +13,20 @@ describe('Component: NovoToast', () => {
             expect(subject).toBeTruthy();
         });
     });
+
+    it('should have a functioning embedded <novo-toast>', () => {
+        // novo-toast is present
+        let subject = element(by.css('.toast-demo novo-toast'));
+        let elm = subject.isPresent().then(() => {
+            // check if novo-toast has appropriate inner content
+            let icon = element(by.css('.toast-icon')).isPresent();
+            let title = element(by.css('.toast-content h5')).isPresent();
+            let message = element(by.css('.toast-content p')).isPresent();
+            if (icon && title && message) {
+                return true;
+            }
+            return false;
+        });
+        expect(elm).toBeTruthy();
+    });
 });
